requirements.txt
requests>=2.31
pandas>=2.0
openpyxl>=3.1
beautifulsoup4>=4.12
lxml>=5.0
python-dotenv>=1.0
streamlit>=1.30
plotly>=5.18
